Extract utilities to 'utils' module
